// src/types/jp_address.d.ts
export interface IJP_Address {
    postal_code?: string;
    prefecture?: string;
    municipality?: string;
    town_area?: string;
    romanization?: string;
    prefecture_hi?: string;
    municipality_hi?: string;
    town_area_hi?: string;
    office_code?: number;
    area_code?: number;
    agency?: string;
    added_at?: Date;
}